refactor: move vega-specific spec to `altair` package
