+++ draco/renderer/__init__.py
@@ -1,4 +1,5 @@
-from .altair_renderer import AltairRenderer
+from draco.renderer.altair.altair_renderer import AltairRenderer
+
 from .base_renderer import BaseRenderer
 
 __all__ = ["BaseRenderer", "AltairRenderer"]
